% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBayesRsquared.R
\name{getBayesRsquared}
\alias{getBayesRsquared}
\title{Get posterior summaries of the Bayesian R squared}
\usage{
getBayesRsquared(
  blockGraph,
  blockList,
  outFilePath,
  outFilePrefix,
  varNames,
  hist = FALSE
)
}
\arguments{
\item{blockGraph}{adjacency matrix for blocks (as input to \code{\link{rHESS_SEM}})
blockGraph should be a square matrix (adjacency for blocks)
columns are LHS of regressions
rows are input RHS of regressions}

\item{blockList}{list of variables in each block (as input to \code{\link{rHESS_SEM}})}

\item{outFilePath}{filepath for RBSEM output files}

\item{outFilePrefix}{file prefix for RBSEM output files (this will be the name of the input data file, stripped of any .txt)}

\item{varNames}{vector of variable names from original input data file}
}
\value{
None
}
\description{
This function loops over all the multivariate regressions, and for each
regression coefficient gets the Bayesian R squared values for each response variable.
Currently this function prints to screen the posterior mean and median R squared
for each response variable. It also produces (optionally) produces histograms 
of the whole posterior R squared for each response. 

The Bayesian R squared is calculated according to Gelman et al. 
The American Statistician, Volume 73, 2019.
}
\examples{
require(utils)
dir.create("tmp")
data(sample_SEM)
write.table(sample_SEM,"tmp/sem_data.txt",row.names = FALSE,col.names = FALSE)
blockL = list( 
  c(9:28),  ## x0 -- block 0
  c(1:5),  ## y1 -- block 1
  c(6:8)  ## y2 -- block 2
)
G = matrix(c( 
  0,1,1,
  0,0,1,
  0,0,0 ), 
  byrow=TRUE,ncol=3,nrow=3)
rBSEM::rHESS_SEM(inFile="tmp/sem_data.txt",blockList = blockL,
         SEMGraph = G,outFilePath="tmp/",nIter=50,method = 0,nChains = 2)
getBayesRsquared(blockList = blockL,blockGraph = G,outFilePath="tmp/",
                outFilePrefix="sem_data",varNames=names(sample_SEM))
unlink("tmp", recursive=TRUE)

}
