% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rHESS_SEM_internal}
\alias{rHESS_SEM_internal}
\title{rHESS_SEM_internal}
\arguments{
\item{inFile}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run}

\item{seed}{pRNG seed}

\item{method}{\deqn{\gamma}{gamma} sampling method, where 0=\deqn{MC^2}{MC^3} and 1=Thompson -sampling-inspired novel method}
}
\description{
Run a simple SEM Bayesian sampler -- internal function
}
