% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBSEM.R
\docType{data}
\name{sample_SEM}
\alias{sample_SEM}
\title{sample synthetic SEM}
\format{A sampler data frame conforming with the software specs 200 obs and 28 variables}
\usage{
data(sample_SEM)
}
\description{
rBSEM -- Bayesian Structural Equation Models Regression
}
\keyword{datasets}
