% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceToArray.R
\name{traceToArray}
\alias{traceToArray}
\title{traceToArray}
\usage{
traceToArray(fileName, nIterations = NULL, nPredictors = NULL)
}
\arguments{
\item{fileName}{path to MCMC trace file}

\item{nIterations}{number of iteration produced by rHESS_SEM}

\item{nPredictors}{number of predictors expected}
}
\description{
Transform the MCMC trace file resulting from calling rHESS_SEM into a manageable array
}
\details{
rBSEM -- Bayesian Structural Equation Models Regression
}
