% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rHESS_SEM}
\alias{rHESS_SEM}
\title{rHESS_SEM}
\usage{
rHESS_SEM(inFile, outFilePath, nIter = 10L, nChains = 1L, seed = 0L,
  method = 0L)
}
\description{
Run a simple SEM Bayesian sampler
}
\examples{
rHESS_SEM("Data/sem_data.txt","Data/",..)

}
