% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rHESS_SEM.R
\name{rHESS_SEM}
\alias{rHESS_SEM}
\title{rHESS_SEM}
\usage{
rHESS_SEM(inFile, blockList, varType = NULL, SEMGraph, outFilePath = "",
  autoAddIntercept = TRUE, gammaInit = "S", nIter, nChains = 1,
  seed = 0, method = 1)
}
\arguments{
\item{inFile}{path to data file}

\item{blockList}{list of blocks in the model}

\item{varType}{variable type for each column in the data file}

\item{SEMGraph}{graph adjacency matrix representing the SEM structure between the blocks}

\item{outFilePath}{path to where the output is to be written}

\item{autoAddIntercept}{should the c++ code automatically add an intercept to every equation?}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run}

\item{seed}{pRNG seed}

\item{method}{\deqn{\gamma}{gamma} sampling method, where 0=\deqn{MC^2}{MC^3} and 1=Thompson -sampling-inspired novel method}
}
\description{
Run a simple SEM Bayesian sampler
}
\details{
rBSEM -- Bayesian Structural Equation Models Regression
}
\examples{
require(utils)
dir.create("tmp")
data(sample_SEM)
write.table(sample_SEM,"tmp/sem_data.txt",row.names = FALSE,col.names = FALSE)
blockL = list( 
  c(9:28),  ## x0 -- block 0
  c(1:5),  ## y1 -- block 1
  c(6:8)  ## y2 -- block 2
)
G = matrix(c( 
  0,1,1,
  0,0,1,
  0,0,0 ), 
  byrow=TRUE,ncol=3,nrow=3)
  rBSEM::rHESS_SEM(inFile="tmp/sem_data.txt",blockList = blockL,
                   SEMGraph = G,outFilePath="tmp/",nIter=50,method = 0,nChains = 2)
unlink("tmp", recursive=TRUE)

}
