% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rHESS_SEM}
\alias{rHESS_SEM}
\title{rHESS_SEM}
\usage{
rHESS_SEM(inFile, outFilePath, nIter = 10L, nChains = 1L, seed = 0L,
  method = 0L)
}
\arguments{
\item{inFile}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}
}
\description{
Run a simple SEM Bayesian sampler
}
\examples{
rHESS_SEM(inFile="Data/sem_data.txt",outFilePath="Data/",nIter=200)

}
