% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rHESS_SEM_internal}
\alias{rHESS_SEM_internal}
\title{rHESS_SEM}
\usage{
rHESS_SEM_internal(inFile, outFilePath, nIter = 10L, nChains = 1L,
  seed = 0L, method = 0L)
}
\arguments{
\item{inFile}{path to data file}

\item{outFilePath}{path to where the output is to be written}

\item{nIter}{number of iterations}

\item{nChains}{number of parallel chains to run}

\item{seed}{pRNG seed}

\item{method}{\deqn{\gamma}{gamma} sampling method, where 0=\deqn{MC^2}{MC^3} and 1=Thompson -sampling-inspired novel method}
}
\description{
Run a simple SEM Bayesian sampler -- internal function
IMPORTANT NOTE: outFilePath must exists, otherwise no output is going to be written.
You can make sure of its existence by using base::dir.create(outFilePath) from R
}
\examples{
dir.create("tmp")
data(sample_SEM)
write.table(sample_sem,"tmp/sem_data.txt",row.names = FALSE,col.names = FALSE)
rHESS_SEM_internal(inFile="tmp/sem_data.txt",outFilePath="tmp/",nIter=200)
unlink("tmp", recursive=TRUE)

}
